refactor: Remove the "SEED" naming convention, it's just a type
