+++ chidian/__init__.py
@@ -4,7 +4,7 @@ from .dict_group import DictGroup
 from .dsl_parser import parse_path_peg as parse_path
 from .mapper import Mapper
 from .partials import ChainableFn, FunctionChain
-from .seeds import DROP, KEEP
+from .types import DROP, KEEP
 
 __all__ = [
     "get",
